new option for mcp servers
This patch adds an option to be able to pass a `verify` option to
MCP servers definition. It is set to true by default which means
that TLS certificate are being checked like they were before the
patch if unset. If set to `false` then the TLS certificate are
not checked.

This option is really useful to have agent zero work with MCP
servers on internal network where it is difficult to have
public and valid certificate.

It will be necessary to update the MCP python module to get access
to the option allowing to pass custom HTTP client.